Remove unreachable segments in case statements
If brackets are used for a `case` statement, the segment with the closing bracket is never "reached", so it llvm-cov shows that as a coverage gap.  :(

It still shows gaps (Xcode highlights with striped red) in the fall-throughs (because we never can fall through), but it does not actually count that as a coverage gap, fortunately.

There are other coverage gaps, but this removes all of this class of switch statement based gaps.