Update windows main
This makes `ku_test` return an `int`, as it is the test `main` function.  It returns the number of test failures, so it is easier to chain a test runner executable with other commands if desired (make may do that).

It also updates kumuwin to properly include the correct header for the test runner mode.  No PC available to me at the moment, so it is untested, but it is already broken, so this can't really make it much worse.